ANP-60 Moved logic to show enable Bluetooth option to BLEManager, fixed bluetoothStatusMessage not updating if Bluetooth status is changed without scanning in meantime
